otp code

+++ accounts/views.py
@@ -1,4 +1,7 @@
 import random
+
+from datetime import timedelta
+from django.utils import timezone
 from django.shortcuts import render, redirect
 from django.views import View
 from django.contrib import messages
@@ -10,9 +13,10 @@ from utils import send_otp_code
 
 class UserRegisterView(View):
     form_class = UserRegistrationForm
+    template_name = 'accounts/register.html'
 
     def get(self, request):
-        return render(request, 'accounts/register.html', {'form': self.form_class})
+        return render(request, self.template_name, {'form': self.form_class})
 
     def post(self, request):
         form = self.form_class = UserRegistrationForm(request.POST)
@@ -31,14 +35,15 @@ class UserRegisterView(View):
             messages.success(request, 'we sent you a code', extra_tags="success")
             return redirect("accounts:verify_code")
 
-        return render(request, 'accounts/register.html', {'form': self.form_class})
+        return render(request, self.template_name, {'form': self.form_class})
 
 
 class UserRegistrationVerifyCodeView(View):
     form_class = VerifyCodeForm
+    template_name = 'accounts/verify.html'
 
     def get(self, request):
-        return render(request, 'accounts/verify.html', {'form': self.form_class})
+        return render(request, self.template_name, {'form': self.form_class})
 
     def post(self, request):
         user_session = request.session["user_registration_info"]
@@ -46,7 +51,8 @@ class UserRegistrationVerifyCodeView(View):
         form = self.form_class(request.POST)
         if form.is_valid():
             cd = form.cleaned_data
-            if cd["code"] == code_instance.code:
+            expiration_time = code_instance.created_at + timedelta(minutes=5)
+            if cd["code"] == code_instance.code and expiration_time < timezone.now():
                 User.objects.create_user(phone_number=user_session["phone_number"], email=user_session["email"]
                                          , full_name=user_session["full_name"], password=user_session["password"])
                 code_instance.delete()
@@ -56,4 +62,4 @@ class UserRegistrationVerifyCodeView(View):
                 messages.error(request, 'wrong code', extra_tags="danger")
                 return redirect("accounts:verify_code")
 
-        return render(request, 'accounts/verify.html', {'form': self.form_class})
+        return render(request, self.template_name, {'form': self.form_class})
